adapt and refactor tests
