+++ tests/test_sailor/test_assetcentral/test_equipment.py
@@ -20,8 +20,8 @@ class TestEquipment:
     def eq_obj(self, make_equipment):
         return make_equipment(equipmentId='D2602147691E463DA91EA2B4C3998C4B', name='testEquipment', location='USA')
 
-    @patch('sailor.assetcentral.equipment._apply_filters_post_request')
-    @patch('sailor.assetcentral.equipment._fetch_data')
+    @patch('sailor._base.apply_filters_post_request')
+    @patch('sailor.assetcentral.equipment._ac_fetch_data')
     def test_find_equipment_indicators_fetch_and_apply(self, mock_fetch, mock_apply, mock_url,
                                                        eq_obj, make_indicator_set):
         object_list = Mock(name='raw_object_list')
@@ -38,7 +38,7 @@ class TestEquipment:
         assert mock_apply.call_args.args[:-1] == (object_list, filter_kwargs, extended_filters)
         assert actual == expected_result
 
-    @patch('sailor.assetcentral.equipment._fetch_data')
+    @patch('sailor.assetcentral.equipment._ac_fetch_data')
     def test_find_failure_modes(self, mock_fetch, mock_config, eq_obj):
         mock_fetch.return_value = [{'ID': 'fm_id1'}, {'ID': 'fm_id2'}]
         expected = 'expected return value is the value returned by the delegate function "find_failure_modes"'
@@ -64,7 +64,7 @@ class TestEquipment:
                                                   equipment_id=eq_obj.id)
             assert actual == expected
 
-    @patch('sailor.assetcentral.location._fetch_data')
+    @patch('sailor.assetcentral.location._ac_fetch_data')
     def test_location_returns_location(self, mock_fetch, mock_config, make_equipment):
         equipment = make_equipment(equipment_id='123', location='Walldorf')
         mock_fetch.return_value = [{'locationId': '456', 'name': 'Walldorf'}]
